improve node identifier documentation and add failure case to `shortId()` (#21407)
## Description

This improves the documentation on the public-facing identifier APIs.
There is now clearer guidance around what the APIs are useful for and
how to use them. The documentation avoids mentioning the "identifer
index" at this time, since it does not yet exist, but when it does we
will update these docs to also highlight the synergy there.

This also adds an error for an edge case - calling `shortId()` when a
node had multiple identifier fields. The previous behavior was to return
_one_ of the identifiers in a compressed form, but it didn't specify
which. Now, it asserts that there is only one identifier field present
on the node.