[Summarization] Update the summarizer / GC APIs on channels to use visitor pattern (#21379)
Follow up from this comment -
https://github.com/microsoft/FluidFramework/pull/21347#discussion_r1633882018

During summarization or attachments, both GC and summarize runs and they
iterate over data stores and channel contexts during summarization. The
logic to filter and validate is duplicated for both which makes it prone
to get diverged.
This PR updates the following set of function to use the
[visitor](https://en.wikipedia.org/wiki/Visitor_pattern) pattern to
iterate over data stores / channel contexts:
- `summarize` and `getGCData`.
- `getAttachSummary` and `getAttachGCData`. 

One logic change in this PR:
- It updates `ChannelCollection::getAttachSummary` to use a visisted set
to ensure that a node is only summarized oncen similar to
`FluidDataStoreRuntime::getAttachSummary`. It earlier used the summary
tree to do this.

#8257