[GC] Fix GC cross compat tests failing for ODSP and AFR (#21443)
## Issue and root-cause

The test "GC summary compatibility tests" has been failing consistently
for ODSP and AFR. The test fails because it does the following:
1. Summarizer 1 generates summary at say seq#100.
2. Summarizer 2 loads and assumes that it will load from summary
seq#100.

This works fine for local server because it will return the above
summary as latest when summarizer 2 loads. However, ODSP and AFR cache
snapshots. So, summarizer 2 actually gets the previous cached summary
which is older than seq#100. Summarizer 2 then receives the ack from
summary seq#100 and it shuts down because its newer than the summary it
knows about. The summary generation from summarizer 2 fails and the test
fails.

## Fix
The fix is that summarizer 2 explcitly loads from the summary generated
by summarizer 1 via the id of the summary.

For ODSP, there is an additional fix required for back-compatibility due
to single-commit summaries. Loaders version 1.x do not support single
commit summaries so when the test runs in this compat mode, ODSP nacks
all summaries. Added logic to skip the tests for loader versions 1.x.




[AB#7948](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7948)