build: Use flub generate assertTags for assert tagging (#21458)
Assert tagging has historically been part of policy-check, but it does
not fit there for many reasons, including that we don't need to run it
as a part of every build, only as part of release prep. This change
switches to `flub generate assertTags`, which will allow us to fully
remove the policy-check handler logic in a future change.

I verified that this command creates the same changes as the existing
command; #21457 demonstrates the results.