Don't send ops during legacy SharedTree summarization (#21459)
## Description

Previously, SharedTree would send ops from the summarizer client during
certain race conditions. The op was redundant and unnecessary, but was
nonetheless submitted. Recently, it has become an error to send to an op
from a summarizer client, so this prevents the sending of the op. No
meaningful change in the behavior of the legacy tree is expected from
this change.

This also updates the legacy tree package's dds stress test command to
include the `node_modules` prefix. Without it, the command will fail
when running `test:stress` locally.