Runtime: Begin to deprecate Compat Details on ContainerRuntime messages (#21355)
This last year we added support to include "compat details" on an op,
with the one feature there being the ability to say "it's ok if you
don't recognize this op type, just ignore it". We added this to try to
speed up early development of features that introduce new op types - GC
in particular.

But upon further reflection, this is a very dubious prospect. It glosses
over a ton of complex cases that end up making little sense. Not worth
any perceived utility.

So we are deprecating it and will soon remove it.