refactor(telemetry-utils): Make `MockLogger`'s `events` property externally read-only (#21470)
There was 1 existing test case that depended on mutability to work
around the fact that the mock logger clears its events after checking
for a match. It accomplishes this by adding an option for whether or not
to clear the events after performing a match check (default: true).

Also fixed a couple of bugs: 
* The early-out check in `matchEventStrict` would cause the events not
to be cleared when they should have been.
* Errors logged by assertion messages would not include the "actual"
event details, because they were cleared before the error was thrown.