build(client): Update typetests' previous version to RC5 and reset overrides (#21543)
Command used to reset the tests and use RC5:

```shell
flub typetests -g client --reset --normalize --exact 2.0.0-rc.5.0.0
```

I also manually removed the `--publicFallback` flag from the `flub
generate typetests` command where it was used. It's no longer needed
because the previous versions now expose the ./internal entrypoint.

### Exceptions

I'll follow up on these later this week after the main branch is
unlocked.

**driver-definitions**: Type tests were mostly busted for this project,
I think due to massive changes in the exported API surface in #21200. I
switched it to test only the _public_ API instead of internal, which
isn't ideal but better than disabling the tests completely. It's not
clear to me if this represents a real customer-impacting break or not.

**file-driver**: This type test required _forward-compatibility_
overrides: VariableDeclaration_FluidFetchReaderFileSnapshotWriter

I think that's OK, but if not we can follow up with a separate change.