Fix nexus uncaught exception during graceful shutdown (#21498)
## Description

During nexus graceful shutdown, we need to stop receiving new
connections. However, the code was throwing uncaught exception as
"res.status is not a function". So, fixing that by using res.writeHead
instead.

Tested it with fluid-chat and verified that the new connections were
receiving 503 during graceful shutdown as expected.

Co-authored-by: Shubhangi Agarwal <shuagarwal@microsoft.com>