improvement(api-markdown-documenter): Simplify default heading text policy (#21535)
A previous update to the library made the default policy for heading
text include alert text for alpha/beta items. While this may be a
reasonable choice for some consumers, it is a fairly opinionated policy,
and probably doesn't belong in the default behavior. Consumers who wish
to emulate this can do so by overriding `getHeadingTextForItem` in their
config.

An additional cause for this change is that the fluid-framework repo is
currently leveraging the `@alpha` release tag to mean something other
than "alpha". As a result, the current API docs published on
fluidframework.com indicate that items are "alpha" when they really
aren't. Changing this default policy is a convenient way to remove
misleading alerts from those docs, while also making it more sensible /
unopinionated.

See updated test collateral for examples of the impact.


[AB#8406](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8406)