Add more AzureClient e2e tests for SharedTree (#21506)
## Description

This adds a smattering of e2e tests for SharedTree to provide additional
coverage of its main features and APIs. These tests are not meant to be
completely exhaustive of every edge case (that is the job of the ST unit
tests), but instead to be a "breadth first" sampling of SharedTree
capabilities to improve upon the (currently very light) e2e test
coverage.

This also refactors the setup code for the two existing SharedTree tests
into a helper function.

TODO: There is a known bug in node identifier encoding that will cause
one of the new tests to fail. This PR must be rerun after the fix goes
in.