Hydrate nodes in the initial tree passed to SharedTree's view initialization (#21565)
## Description

Previously, nodes in the initial tree were not hydrated (as they are
when inserting content in general). This meant that users holding on to
references to the nodes passed to initialization would not find them to
be the same as the nodes that are read out of the tree after
initialization. This PR updates the behavior to be the same as when
inserting nodes - the inserted nodes can be used after the insertion.