fix(docs): Map `@alpha` tags to "Legacy" alerts in API documentation (#21525)
In our repo, we are currently using the `@alpha` release tag to mean
something different than "alpha". We currently map it to a concept we're
calling "legacy". This PR updates the generated API docs to use "Legacy"
terminology in various tables in the API docs, instead of "Alpha" to
avoid any customer confusion.

In the future, we plan on reclaiming `@alpha` for its traditional
meaning, and using a `@legacy` tag for the special concept. The
configurations changed here will be need to changed again at that time.

Before

![image](https://github.com/microsoft/FluidFramework/assets/54606601/95f14088-5620-4ad6-834a-acdda73ce274)

After

![image](https://github.com/microsoft/FluidFramework/assets/54606601/29d039fa-c448-41f1-9beb-c5bcf4d7c002)


[AB#8406](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8406)