Patch (#21568): Remove O(n) slowdown from event listeners  (#21573)
## Description

This is a cherry-pick of #21568 

There are currently scenarios in which hundreds of thousands of event
subscriptions can occur on nodes in a SharedTree (that is a separate bug
that is also being fixed) and that case, the linear walk of the event
registrations that is required to validate that an event listener is not
a duplicate is prohibitively expensive. This removes the error, and
slightly relaxes the contract to be that duplicate event listeners cause
undefined behavior, rather than an error.