server: cleanup orderer connection on socket disconnect (#21528)
## Description

We know that `Nexus` Lambda has a memory leak. #5428 from @GaryWilber
added `KafkaOrdererFactory.delete` to allow us to prevent memory leaks,
but it was unused in Routerlicious.

The following changes are included in this PR so that we actually use
the `delete` function:

- Change `KafkaOrdererFactory.delete` to also `close` the `Orderer`
after deletion.
- Begin tracking count of orderer connections for a given
tenantId/documentId in `IOrdererManager`
- Expose `IOrdererManager.removeOrderer` method which is called in
`Nexus` after disconnecting from the OrdererConnection to allow
`IOrdererManager` to properly cleanup its internal Map(s)
- Orderer connections are cleaned up on a 60 second delay to reduce
likelihood of having to immediately re-setup connection on network
reconnect blip or Read->Write connection change

## Breaking Changes

- Add new `IOrdererManager.removeOrderer` public method
- Implement `IOrdererManager.removeOrderer` in `LocalOrdererManager`
- Implement `IOrdererManager.removeOrderer` in `OrdererManager`
- Change `KafkaOrdererFactory.delete` to return `Promise<void>` from
`void`

## Reviewer Guidance

The real meat of this change is in the `OrdererManager` class. Please
focus your review there if time is limited.