[bump] client: 2.0.0-rc.5.0.2 => 2.0.0 (exact) (#21592)
Bumped client from 2.0.0-rc.5.0.0 to 2.0.0.

Command used:

```shell
flub bump client --exact 2.0.0
```

Includes a manual code change in
packages/test/test-end-to-end-tests/src/test/deRehydrateContainerTests.spec.ts
to work around test failures. The intent of that bit of code is to draw
a cut line on whether to use the old or new API for an API that was
changed in 2.0.0-internal.8.0.0. The extra check ensures that 2.0 dev/PR
builds - e.g. `2.0.0-12345` - are considered as "current" versions.