build(client): Include markdown-magic and changelog-generator-wrapper in client release group (#21393)
The markdown-magic and changelog-generator-wrapper packages are both
private packages that we use in various places within the repo. We have
been using `file:` links to these packages in the root workspace, but
that causes `pnpm i` to always trigger resolution, which not only wastes
time, but produces unnecessary lockfile changes in some cases.

I left the markdown-magic package where it is to minimize the change,
since it has a build pipeline and we continue to use `file:`
dependencies in other places like docs.

changelog-generator-wrapper, on the other hand, was moved to
packages/tools. We can do the same for markdown-magic once its build
pipeline is disabled/removed.

Including these packages in the root workspace caused the following
changes:

- Some dependency versions were changed to match those used in the
client release group.
- The version of the packages was updated to match the client release
group version. These packages are not published so the version should be
irrelevant.