Put Memory Blob Storage Under a Flag (#21612)
We recently made changes to how detached blob storage works, and enabled
a default in-memory blob storage such that blobs would be supported by
default in detached containers.

However, when blobs are in a detached container, it changes the
container attachment flow, specifically, an empty file is created, the
blobs are uploaded, and then the summary is uploaded. Due to this
differing flow some drivers, like odsp, create files with a `.tmp`
extension appended to the file name, this is meant to signify that that
file could be incomplete. For example, the summary could fail to upload
after blobs have been uploaded, and this will not be a valid fluid file.
it is expected that the application will rename the file after
attachment completes if the `.tmp` extension is not desired. If an
application does not expect the `.tmp` extension this can create issues,
so for now, we will put this support under a flag, that must be
explicitly enabled `"Fluid.Container.MemoryBlobStorageEnabled" = true`
so that apps must opt into this behavior.

related to #21144

Co-authored-by: Tony Murphy <anthonm@microsoft.com>