Stop using short data store IDs in detached container (#21613)
Currently, data stores created in detached container always get a short
ID via id compressor. We have a bug in one of our customer's app where a
short data store ID created is `[` but in a downloaded snapshot, it is
converted to its ASCII equivalent `%5B` in certain conditions. So, when
an op comes for this data store with id `[`, containers loaded with this
snapshot cannot find the data store.

We haven't yet identified what causes the id to change from `[` ->
`%5B`. However, we know that it is caused by short IDs generated in
detached container.

To mitigate the issue for customers, this PR disables the generation of
short IDs for data stores and DDSes. Added tests that validate that we
don't generate short IDs for data stores and DDSes in detached or
attached state.

(cherry picked from commit 1c00d7e38b430c6c9df201536a95b2723a75c3bc)