build(client): generate legacy API report (#21634)
client-utils is in the `@fluid-internal` scope and thus does not have
trimmed exports. But it currently support legacy APIs for typed event
emitter.

Setup a non-production legacy "roll up" file for both the browser and
Node.js entrypoints. Generate API reports and perform linting like
`@fluidframework` scoped packages with /legacy, but using the faux "roll
ups".

Additionally now verify API reports for browser and Node.js entrypoints
are identical.

Note: `api` task and `api-extractor` named scripts are not used as in
`@fluidframework` scoped packages as these scripts are not part of
production API. They are only used for docs and linting which have task
dependencies defined.