Set up fetch-tool for usage of a different app registration (#21608)
## Description

Adjusts the `IPublicClientConfig` that fetch-tool uses to obtain a
distinct clientId from the environment. This sets us up better for using
a less-permissioned app registration (fetch-tool only doesn't need write
access to graph). I haven't yet created this new app registration, so
this is for now a no-op on the registration we're using.

I've also updated `odspsnapshotfetch-perftestapp` to use the same
registration, as its intent closely matches fetch tool (test perf of
fetching snapshot on a document the signed in user has access to).

For some more context on the direction we're moving in for managing
different flows against odsp, see also #21591 or changes to internal
documentation on infrastructure.

Co-authored-by: Abram Sanderson <absander@microsoft.com>