[main>rc4] Fix comparison of 2.0 releases to RC builds (#21651)
Ports #21641 to RC4.

With the 2.0.0 release, we now have semver releases that are compared
with RC and internal build numbers. I have fixed the logic in
version-tools to correctly treat RC -> 2.0 as a major bump.

I also updated the generate:buildVersion command to use the list of tags
passed in for all logic. There was some code that checked for existing
tags and it always used the in-repo tag list. Now, if a list of tags is
passed in, it is used for all logic. The repo tags are only consulted
when none are passed in.