Added tests to validate that DDS created with specfial characted in ID works correctly (#21657)
Follow up to https://github.com/microsoft/FluidFramework/pull/21625.

This PR adds tests to validate that when a DDS is created with special
characted "[" in it's ID, everything works as expected. In particular,
during summary upload the ID is URI encoded correctly and during
snapshot download it should be URI decoded correctly. Currently, we have
a bug where the decoding doesn't happen when snapshot is downloaded.
The test does the following:
- Creates a DDS with `[` in its id in a detached container.
- Attaches the container.
- Loads a container with the above snapshot.
- Resets the document service factory so that it doesn't load the
uploaded snapshot from the cache.
- Loads the DDS and validates that it fails today. When we have a fix,
the test will start failing and the validation should be updated.


[AB#7045](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7045)