Periodic snapshot refresh (#21504)
This change triggers fetching the latest snapshot in the
serializedStateManager after a timeout established by config provider.
Once that is complete, ssm will try to update the base snapshot if the
already given conditions are met. Such timer is initialized in three
different scenarios:
1. While setting the initial snapshot while attaching the container.
2. While loading a container with no pending state.
3. Once we updated the snapshot for a subsequent refresh.

The scenario in which we load from a container with pending state is
already covered by current code.

Here are some scenarios this change is covering, explaining its
behavior:

1. Start fetching once timer is finished, fetching succeeds, conditions
to refresh are met, base snapshot is updated, and a timer is restarted
for the next attempt. On the next attempt the same procedure is
followed.
2. Start fetching once timer is finished, fetching succeeds, conditions
to refresh are not met, wait for all conditions to be met and attempt
again via saved event, snapshot is updated.
3. Start fetching once timer is finished, fetching succeeds, conditions
to refresh are met, snapshot is not updated due to being too old to
refresh, timer restarts.
4. Start fetching once timer is finished, fetching does not succeed,
attempt again until it succeeds.

Current conditions to update the snapshot should be sufficient since the
most complicated case (while loading with pending state) is already
covered.