Offline: Plumb clientSequenceNumber from loader submit APIs to PendingStateManager again (#21575)
It's a partial revert of #13994.

The Container's various "submit" functions all return a client sequence
number (CSN) which is assigned sequentially when the loader layer sends
the message to the ordering service. For a batch, the CSN returned is
the CSN of the last message in the batch, including chunks sent during
chunking (if applicable).

We take this and convert it to the CSN of the start of the batch. In a later change,
this will be used to form a new Batch ID for detecting duplicate changes
across container forks, e.g. when improperly rehydrating from a
serialized container.