Fix bug for No context for Op (#21680)
## Breaking Changes

[ADO item
link](https://dev.azure.com/fluidframework/internal/_workitems/edit/7045)

The issue was that we were encoding the ids of datastores/ddses when
uploading the summary while on the snapshot download path for binary
snapshot we were decoding them. Previously, when we were not using short
ids, the encoded ids were same as original ids as they did not contain
special characters, so we never faced this issue before. With the usage
of short ids, we discovered this long-hidden bug for ODSP.
Also, enabled the e2e tests which were written to reproduce the
scenario.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>