[Summarization] Added tests for scenarios where summarizer receives newer summary ack (#21711)
There can be scenarios such as DB rollback in server, where the latest
snapshot is lost but the corresponding summary ack is still present.
Currentlty, if a summarizer receives an ack that is newer that the
summary it's tracking, it will fetch latest snapshot and close. So, in
the above scenario, the summarizer would be stuck in loop fetching
snapshot and closing corrupting the document. There is no way out of
this state.

I have added tests for the following scenarios which includes the one
described above:
- A summarizer receives a summary ack that is newer and it fetches the
corresponding snapshot.
- A summarizer receives a summary ack that is newer but it fetches an
older snapshot.

Once we have a fix for the bug described above, we can use the above
tests to validate it.

There are couple other changes:
- Combined tests in
`summarizeWithOutOfOrderDataStoreRealization.spec.ts` and
`summarizationEdgeCases.spec.ts` as they were related.
- Fixed "Fluid.ContainerRuntime.Test.CloseSummarizerDelayOverrideMs"
setting to done before each test and reset after.


[AB#8334](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8334)