Offline: Provide incoming batch's starting CSN to the PendingStateManager (#21714)
On submit, the PendingStateManager stores the batch's starting CSN along
with each message. We need the corresponding information for incoming
messages too, for two reasons:

1. We can assert that they match as we process incoming acks for local
messages
2. We need it to compute batch ID, when checking for the same batch
coming from a forked container.

Today, two things happening in RemoteMessageProcessor keep the correct
info from flowing to the PSM:

1. Grouped Batches overwrite clientSequenceNumber with the index into
the batch
2. Without Grouped Batching, batch messages are processed one-by-one, so
the context of the start of the batch is not available when later
messages are sent to PSM.

Here's the fix - the RemoteMessageProcessor has the full context and can
return the batch's starting CSN with each message, irrespective of
grouped batching, compression, chunking, etc.