Refactor BlobManager Snapshot and Summary Code (#21729)
This change is purely about refactoring to make the code easier to
modify later. The change moves all the snapshot and summary
implementation for the blob manager into a single file and lays out a
pattern for versioning. This change also moves to free functions and
removes statics where possible as they are more bundle efficient than
members, and are easier to test and maintain.