fix(version-tools): Fix scheme detection and add more test cases (#21710)
In #21641 I tried to fix version comparisons between semver and RC
prereleasse builds. Unfortunately, the fix was flawed because the scheme
detection was not run on the input version, but rather the "internal
version" in the case of internal/RC builds. This caused patch releases
to get identified as major releases in the RC pipelines, which in turn
causes GH releases to be created incorrectly.

I corrected the code by running schema detection earlier in the
function, and I added several test cases around RC versions to ensure
the bug was addressed.