[main > client/2.0]: RemoteDatastoreContext should handle snapshot with older loader (#21755)
## Description

RemoteDatastoreContext should handle snapshot with groupId with older
loader. So, in case where we fetch the snapshot using old loader and
does not have ISnapshot which contains blobContents, then that means we
must have fetched the full snapshot using older format, so there is no
need to evaluate in that case whether to fetch the snapshot or not.

Enable the test as it passes with odsp.