feat(build-cli): Add release:prepare command (#16686)
One of the valuable things the `release` command does is run a series of
checks to make sure the branch is ready to serve as the base of a
release branch. These checks may require changes, such as tagging
asserts or generating changelogs. These changes need to be merged before
the release branches are created to ensure the release branches have
those changes.

However, the release command runs these checks and immediately exits if
one fails, and it also makes local changes when needed. It is not really
possible to just run the checks and get their results without making
changes.

This PR implements a new `release prepare` command that runs these
checks without making changes and reports the results, along with a
command that may resolve the problem.

### Implementation notes

I considered using the StateMachineCommand as a base and replicating the
checks exactly as they're done in the `release` command, but that
pattern is not as great as I once thought. It would require creating a
new state machine, duplicating pieces of the release one, and it
intuitively feels too complicated for this use. It's arguably too
complicated for its current use in `release`, and this is a much simpler
command.

I also considered adding a new flag to the release command, but it's
already so complex that adding a new flag multiplies it that much more.

Instead, I copied and refactored the core of the existing checks into
new functions that require fewer arguments and return more useful data
than the state machine handler functions.

---------

Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>