build(client): Update build-tools deps to 0.40.0 (#21783)
Updated the following:

  client (release group)

Dependencies on build-tools updated:

  @fluid-tools/build-cli: 0.40.0
  @fluidframework/build-tools: 0.40.0
  @fluidframework/bundle-size-tools: 0.40.0
  @fluid-tools/version-tools: 0.40.0

## Automated changes

This update caused two major changes:

- type tests are now more compact and several bugs have been addressed.
- Some api reports have comment-only changes.

## Manual changes

I also noticed that odsp-doclib-utils had a lint config that did not
extend from "prettier," so some lint rules that shouldn't be running
were getting triggered on the type tests. Extending the prettier config
resolved the problem.

Finally, I added an exclusion of the npm-metadata-and-sorting policy for
the root package.json because of
[AB#8640](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8640),
which I'll resolve later.