[GC] Fixed GC tests that were failing for ODSP after moving to single commit summaries (#21790)
## Bug
The end-to-end tests in file `gcTreeSummaryHandles.spec.ts` have been
failing for ODSP ever since single-commit summaries were enabled. The
reason is that the test intercepted the summary uploaded to driver and
inspected it to find GC summary tree. However, with single commit
summaries, the summary tree structure changed - the container runtime's
summary tree was added under ".app" sub-tree. This was causing the tests
to fail consistently.

## Fix
Updated the tests to use summarize on demand which returns the container
runtime's summary tree. The tests were hacking around to upload summary
and inspect the tree. Removed all of that code and made it much simpler.


[AB#781](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/781)

[AB#6036](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6036)