Remove outdated node hydration regression test (#21806)
## Description

This test was protecting against a scenario that could arise only in
previous versions of the code. Originally, proxies created in the
initial tree were not hydrated, and there was at one point a bug where
SharedTree was attempting to hydrate them, but was doing so incorrectly.
The bug was fixed and this test regression test was added. However,
proxy hydration has since been improved so that the proxies in the
initial tree _are_ hydrated. Therefore, this class of bug no longer
exists, and this test is unnecessary.