Offline: Add batchId to batch metadata on resubmit (#21767)
Batch ID is the key piece of information needed to detect/prevent
duplication of changes due to a forked container.

We compute Batch ID as a tuple of original client ID and original Client
Sequence number (of the batch start). We only stamp the batchId on batch
metadata (of first message in the batch) on resubmit - if a batch is
missing batchId we will be able to compute it from the client ID and
Client Sequence Number.

The next PR will be more interesting - comparing batch IDs in
PendingStateManager as incoming messages are processed. This just sets
it up by adding batch ID to batch metadata on resubmit.

## Breaking Changes

This PR breaks the format of the stashed pending local state.
Back-compat support for reading old-format stashes is included. New
format is backwards compatible (additive). This is slightly superfluous
anyway since there is no production usage of the stashed schema.