[Summarization] Added README for summarization which provides an overview of summaries (#21808)
This document provides a conceptual overview of summarization without
going into a lot of technical or implementation details. It describes
what summaries are, how / when they are generated and what do they look
like. The goal is for this to be an entry point into summarization for
users and developers.
It will also replace the summarizer doc on fluidframework.com -
https://fluidframework.com/docs/concepts/summarizer and anyone reading
this should understand summaries at a very high level.
It omits some advanced / implementation content from the linked document
which are not really part of an overview document. For example, the
details about Summary Manager, Running Summarizer, etc. are not relevant
to someone show wants to just know what summaries are.