Re-enable data virtualization in stress tests (#21864)
[AB#8686](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8686)

Re-enable data virtualization in stress tests. I found that the bug was
not exactly related to virtualized datastores.
https://github.com/microsoft/FluidFramework/pull/21851 What was causing
the issue was creating a shared counter dds as a child to the virtual
datastores. This would cause offline attach ops to hang. I removed this,
so this should now run to completion.