Remove node key index and node key field from flex node (#21868)
## Description

The node key index is not currently used in any practical way, planned
index work completely supersedes it, and the code that is there is not
valuable to keep around. This PR removes it and related test files. It
also removes the `localNodeKey` and related field from the flex tree
interfaces since they are unused.