Fix offline attach ops with child DDSes (#21851)
[AB#8854](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8854)

There was a particularly hard bug that was causing this specific test to
hang. I looked into why we didn't originally catch this bug in the
onset, and it seems that only when we call handle.get during another
handle.get during attach flow do we have this issue.

The repro steps are the following
- Have a container go offline
- Create a `DataObject` with a child dds
- In the `hasInitialized` function of that DDS, load the child dds
- Serialize the container
- load a new container from the serialized state
- The new load hangs.

What happens is that while we are applying the stashed attach op, the
hasInitialized function is running as we are trying to get the entry
point. The entrypoint is a local handle, but the child handle is a
`RemoteFluidObjectHandle`. Thus it calls back into the runtime to get
the child DDS. The problem here is that we hadn't yet put the
`LocalFluidDataStoreContext` we created for applying the stashed op for
our `DataObject` in the list of bound contexts. Thus, we need to do so.

There may be more work to be done here, but this is the fix that gets
around the issue for now.

---------

Co-authored-by: Tony Murphy <anthony.murphy@microsoft.com>