[GC] Added telemetry when an unreferenced data store is realized during summarization (#21860)
Recently, we have seen cases where GC deletes a data store which is
loaded in summarizer but it's not loaded in regular clients. Ideally,
this should never happen because there should be no way to load an
unreferenced data store.
However, the theory is that these data stores have trailing ops that
were sent after they were unreferenced. A summarizer loads after sweep
timeout and realizes the data store due to these trailing ops and then
deletes it. This PR makes following changes to help identify these
scenarios:
- Added telemetry when an unreferenced data store is realized during
summarization. It adds properties such as number of trailing ops and if
fullTree was enabled.
- Changed the `Context is deleted` errors to be data processing errors
instead of data corruption errors because the document is not corrupted
in these cases.
  

[AB#8802](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8802)

Here's a rough sketch Mark made of the scenario, in case it helps (added
here by Mark)

![image](https://github.com/user-attachments/assets/19f4582b-6055-4dc3-b418-a1a1bd411646)